Added BG, Asteroid, Asteroid Spawner, and Health UI
Added  in a BG because I was sick of grey. Started work on an asteroid, along with a spawner that takes a list of positions, randomizes every frame, and spawns it at one. Not happy with the Asteroid movement yet. Dynamic Health UI that responds to both taking and healing damage also functional.